# [Frontend] decrease import time of vllm.multimodal