by changing some modules in `vllm/multimodal` to lazily import expensive modules like `transformers` or only importing them for type checkers when not used during runtime.

contributes to #14924

## `python -c 'import vllm'`

seems slightly faster

### before (main branch commit 302f3aca7ea3f57842881cb2ae0062c19ad24758)

```shell
$ hyperfine 'python -c "import vllm"' --warmup 3 --runs 100 --export-markdown
 out.md
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      9.367 s ±  0.215 s    [User: 8.951 s, System: 2.028 s]
  Range (min … max):    8.931 s … 10.013 s    100 runs
```
 
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `python -c "import vllm"` | 9.367 ± 0.215 | 8.931 | 10.013 | 1.00 |

### after (my PR commit de28f4f933760b7b53aca164ac8c2d7b5256bf11)

```shell
$ hyperfine 'python -c "import vllm"' --warmup 3 --runs 100 --export-markdown out.md
Benchmark 1: python -c "import vllm"
  Time (mean ± σ):      9.196 s ±  0.373 s    [User: 8.758 s, System: 2.065 s]
  Range (min … max):    8.837 s … 12.306 s    100 runs
```

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `python -c "import vllm"` | 9.196 ± 0.373 | 8.837 | 12.306 | 1.00 |