@davidxia 2025-05-13
> not in the scope of this PR, but ideally we want to reduce this 8s as much as possible with lazy load

what's "8s"?

> can you test out if multimodal request still works?

Do you have an example I should run?