@aarnphm 2025-05-13
> what's "8s"?

from your hyperfine run, especially `User: 8.758 s`. This is just a reference for notes, that's all.

> Do you have an example I should run?

There are a few examples in `vllm/examples` [here](https://github.com/vllm-project/vllm/blob/main/examples/online_serving/openai_chat_completion_client_for_multimodal.py)