@davidxia 2025-05-13
> > what's "8s"?
> 
> from your hyperfine run, especially `User: 8.758 s`. This is just a reference for notes, that's all.
ah that's right 😅 

> > Do you have an example I should run?
> 
> There are a few examples in `vllm/examples` [here](https://github.com/vllm-project/vllm/blob/main/examples/online_serving/openai_chat_completion_client_for_multimodal.py)

thanks, I tried that earlier on a CPU platform with `vllm serve llava-hf/llava-1.5-7b-hf`. But running that file caused the server to crash after sending a 500 response. Do you know if that example only works on GPU platforms and isn't supported with CPU?