@aarnphm 2025-05-13
Ah, let me perfor a quick test then if you don't have access to GPU