@davidxia 2025-05-13
> > Ah, let me perfor a quick test then if you don't have access to GPU
> 
> This works with phi 3.5 vision. You can use the diff here
> 
> [diff.patch](https://github.com/user-attachments/files/20180794/diff.patch)

thanks, applied your patch