@davidxia 2025-05-13
> Thanks for taking the piecewise approach. This will be easier to review and merge.

Of course! LazyLoaded PIL.Image in `vllm/multimodal/parse.py`. Ready for another review.