@hmellor 2025-05-13
> @hmellor from the readthedocs logs it seem to build succesfully? do you know if there are any issue with this?

RTD treats warnings as errors:

```
/home/docs/checkouts/readthedocs.org/user_builds/vllm/checkouts/18031/docs/source/api/summary.md:88: WARNING: Could not find vllm.multimodal.inputs.MultiModalDataDict [autodoc2.missing]
/home/docs/checkouts/readthedocs.org/user_builds/vllm/checkouts/18031/docs/source/api/summary.md:95: WARNING: Could not find vllm.multimodal.inputs.NestedTensors [autodoc2.missing]
```