@aarnphm 2025-05-13
ah I see.

@davidxia can you move the docstring in TYPE_CHECKING down to the else block instead? Thanks.