@davidxia 2025-05-13
@aarnphm @hmellor I'm trying fix the sphinx warnings. I tried copying the same `MultiModalDataDict` docstring in the `if TYPE_CHECKING` into the `else`. But when I run `python -m sphinx -T -W --keep-going -b html -d _build/doctrees -D language=en . out/html` locally I still see the same warnings. Any ideas?