@aarnphm 2025-05-14
> `MultiModalDataDict` docstring in the `if TYPE_CHECKING` into the `else`. But when I run `python -m sphinx -T -W --keep-going -b html -d _build/doctrees -D language=en . out/html` locally I still see the same warnings. Any ideas?

probably better to keep previous change, but update the annotations to string instead

i.e: 

```python
if TYPE_CHECKING:
  import torch
 
HfImageItem: TypeAlias = Union[Image, np.ndarray, "torch.Tensor"]
"""docstring as before"""
```